requirements.txt
streamlit==1.28.0
Pillow==10.0.0
pytesseract==0.3.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.15
numpy==1.24.3
matplotlib==3.7.1
plotly==5.16.1
scikit-learn==1.3.0
transformers==4.30.0
beautifulsoup4==4.12.2
requests==2.31.0
python-magic==0.4.27
tqdm==4.65.0
PyPDF2==3.0.1
pdfplumber==0.9.0